refactor(errors): Use a single error type again
Really, the error I designed works well both for compiling and
interpeting and it isn't worth the hassle of dividing the two.
