+++ src/compiler/mod.rs
@@ -1,5 +1,4 @@
 mod block;
-mod error;
 mod include;
 mod lexer;
 mod options;
@@ -7,8 +6,9 @@ mod parser;
 mod tag;
 mod token;
 
+pub use super::error::{Result, Error, ResultLiquidChainExt, ResultLiquidExt};
+
 pub use self::block::{ParseBlock, ParseBlockClone, BoxedBlockParser, FnParseBlock};
-pub use self::error::{Result, ResultCompilerChainExt, ResultCompilerExt, Trace, CompilerError};
 pub use self::include::{Include, IncludeClone, NullInclude, FilesystemInclude};
 pub use self::lexer::{Element, tokenize};
 pub use self::options::LiquidOptions;
