+++ src/tags/assign_tag.rs
@@ -1,4 +1,4 @@
-use error::Error;
+use error::Result;
 
 use interpreter::Context;
 use interpreter::Output;
@@ -6,7 +6,6 @@ use interpreter::Renderable;
 use compiler::LiquidOptions;
 use compiler::Token;
 use compiler::{parse_output, expect, unexpected_token_error};
-use compiler::CompilerError;
 
 #[derive(Clone, Debug)]
 struct Assign {
@@ -15,7 +14,7 @@ struct Assign {
 }
 
 impl Renderable for Assign {
-    fn render(&self, context: &mut Context) -> Result<Option<String>, Error> {
+    fn render(&self, context: &mut Context) -> Result<Option<String>> {
         let value = self.src.apply_filters(context)?;
         context.set_global_val(&self.dst, value);
         Ok(None)
@@ -25,7 +24,7 @@ impl Renderable for Assign {
 pub fn assign_tag(_tag_name: &str,
                   arguments: &[Token],
                   _options: &LiquidOptions)
-                  -> Result<Box<Renderable>, CompilerError> {
+                  -> Result<Box<Renderable>> {
     let mut args = arguments.iter();
     let dst = match args.next() {
         Some(&Token::Identifier(ref id)) => id.clone(),
