+++ src/tags/raw_block.rs
@@ -1,11 +1,10 @@
-use error::Error;
+use error::Result;
 
 use interpreter::Context;
 use interpreter::Renderable;
 use compiler::Element;
 use compiler::LiquidOptions;
 use compiler::Token;
-use compiler::CompilerError;
 
 #[derive(Clone, Debug)]
 struct RawT {
@@ -13,7 +12,7 @@ struct RawT {
 }
 
 impl Renderable for RawT {
-    fn render(&self, _context: &mut Context) -> Result<Option<String>, Error> {
+    fn render(&self, _context: &mut Context) -> Result<Option<String>> {
         Ok(Some(self.content.to_owned()))
     }
 }
@@ -22,7 +21,7 @@ pub fn raw_block(_tag_name: &str,
                  _arguments: &[Token],
                  tokens: &[Element],
                  _options: &LiquidOptions)
-                 -> Result<Box<Renderable>, CompilerError> {
+                 -> Result<Box<Renderable>> {
     let content = tokens.iter().fold("".to_owned(), |a, b| {
         a +
         match *b {
